requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==3.3.0
tox==2.7.0
coverage==4.4
Sphinx==1.5.5
cryptography==1.8.1
PyYAML==3.12
sphinxcontrib-napoleon==0.6.1
ipython==5.3.0 # pyup: ignore
#flake8-pep257==1.0.5
pytest==3.0.7
pep8==1.7.0
yapf==0.16.1
isort==4.2.5
